refs #35894

+++ portlets/activity/src/main/java/com/aimluck/eip/activity/ActivityNotificationFormData.java
@@ -170,13 +170,17 @@ public class ActivityNotificationFormData extends ALAbstractFormData {
     }
     if (ALReminderService.isEnabled()) {
       if (defaultItem != null) {
-        reminder_flag.setValue("T");
-        List<ReminderNotifyType> list = defaultItem.getNotifyType();
-        if (list.contains(ReminderNotifyType.MAIL)) {
-          notify_type_mail.setValue("TRUE");
+        if (defaultItem.isEnabled()) {
+          reminder_flag.setValue("T");
         }
-        if (list.contains(ReminderNotifyType.MESSAGE)) {
-          notify_type_message.setValue("TRUE");
+        List<ReminderNotifyType> list = defaultItem.getNotifyType();
+        if (list != null && list.size() > 0) {
+          if (list.contains(ReminderNotifyType.MAIL)) {
+            notify_type_mail.setValue("TRUE");
+          }
+          if (list.contains(ReminderNotifyType.MESSAGE)) {
+            notify_type_message.setValue("TRUE");
+          }
         }
         notify_timing.setValue(Long.valueOf(defaultItem.getNotifyTiming()));
       }
@@ -242,30 +246,38 @@ public class ActivityNotificationFormData extends ALAbstractFormData {
     }
 
     if (ALReminderService.isEnabled()) {
-      if (reminder_flag.getValue().equals("T")) {
-        if (defaultItem == null) {
-          defaultItem = new ALReminderDefaultItem();
-          defaultItem.setOrgId(orgId);
-          defaultItem.setUserId(loginUser.getName().getValue());
-          defaultItem.setEnabled(true);
-          defaultItem.setCategory(ReminderCategory.SCHEDULE);
+      try {
+        if (reminder_flag.getValue() != null
+          && reminder_flag.getValue().equals("T")) {
+          if (defaultItem == null) {
+            defaultItem = new ALReminderDefaultItem();
+            defaultItem.setOrgId(orgId);
+            defaultItem.setUserId(loginUser.getName().getValue());
+            defaultItem.setEnabled(true);
+            defaultItem.setCategory(ReminderCategory.SCHEDULE);
+          }
+          ArrayList<ReminderNotifyType> list =
+            new ArrayList<ReminderNotifyType>();
+          if (notify_type_mail.getValue() != null
+            && notify_type_mail.getValue().equals("TRUE")) {
+            list.add(ReminderNotifyType.MAIL);
+          }
+          if (notify_type_message.getValue() != null
+            && notify_type_message.getValue().equals("TRUE")) {
+            list.add(ReminderNotifyType.MESSAGE);
+          }
+          defaultItem.setNotifyType(list);
+          defaultItem.setNotifyTiming(notify_timing.getValueWithInt());
+          ALReminderService.updateDefault(defaultItem);
+        } else {
+          ALReminderService.removeDefault(
+            orgId,
+            loginUser.getName().getValue(),
+            ReminderCategory.SCHEDULE);
         }
-        ArrayList<ReminderNotifyType> list =
-          new ArrayList<ReminderNotifyType>();
-        if (notify_type_mail.getValue().equals("TRUE")) {
-          list.add(ReminderNotifyType.MAIL);
-        }
-        if (notify_type_message.getValue().equals("TRUE")) {
-          list.add(ReminderNotifyType.MESSAGE);
-        }
-        defaultItem.setNotifyTiming(notify_timing.getValueWithInt());
-        defaultItem.setNotifyType(list);
-        ALReminderService.updateDefault(defaultItem);
-      } else {
-        ALReminderService.removeDefault(
-          orgId,
-          loginUser.getName().getValue(),
-          ReminderCategory.SCHEDULE);
+      } catch (Exception e) {
+        logger.error("Unable to save ReminderService ", e);
+        return false;
       }
     }
